Fixes Review Comments and seperated out debug_envgetter.py
