+++ src/databricks/labs/remorph/connections/env_getter.py
@@ -1,24 +1,13 @@
 import os
-import json
-import logging
 
 
 class EnvGetter:
-    def __init__(self, is_debug: bool = False):
-        self.env = self._get_debug_env() if is_debug else dict(os.environ)
+    """Standardised inorder to support testing Capabilities, check debug_envgetter.py"""
+
+    def __init__(self):
+        self.env = dict(os.environ)
 
     def get(self, key: str) -> str:
         if key in self.env:
             return self.env[key]
         raise KeyError(f"not in env: {key}")
-
-    def _get_debug_env(self) -> dict:
-        try:
-            debug_env_file = f"{os.path.expanduser('~')}/.databricks/debug-env.json"
-            with open(debug_env_file, 'r', encoding='utf-8') as file:
-                contents = file.read()
-            logging.debug(f"Found debug env file: {debug_env_file}")
-            raw = json.loads(contents)
-            return raw.get("ucws", {})
-        except FileNotFoundError:
-            return dict(os.environ)
